:recycle: Separated types into component files
