+++ ice_station_zebra/types/enums.py
@@ -1,0 +1,15 @@
+from enum import IntEnum, StrEnum
+
+
+class BetaSchedule(StrEnum):
+    """Enum for diffusion beta schedule types."""
+
+    LINEAR = "linear"
+    COSINE = "cosine"
+
+
+class TensorDimensions(IntEnum):
+    """Enum for tensor dimensions."""
+
+    THW = 3  # Time, Height, Width
+    BTCHW = 5  # Batch, Time, Channels, Height, Width
